requirements.txt
Flask==3.0.0
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
pyodbc==5.0.1
python-dotenv==1.0.0
Pillow==10.1.0
gunicorn==21.2.0
